feat: Cadastro de produtos

+++ src/app/hooks/product.tsx
@@ -1,28 +1,60 @@
 "use client";
 import { HookChildrenProp } from "../types/children";
-import { Product, ProductHookProps } from "../types/product";
+import { Product, ProductHookProps, ProductSaveData } from "../types/product";
 import { createContext, useContext, useState } from "react";
 import { api } from "@/utils/api";
-import { AxiosRequestConfig } from "axios";
+import { AxiosError, AxiosRequestConfig } from "axios";
+import { useLoading } from "./loading";
+import { useRouter } from "next/navigation";
 
 const ProductContext = createContext({} as ProductHookProps);
 
 export const ProductHook = ({ children }: HookChildrenProp) => {
   const [products, setProducts] = useState([] as Product[]);
+  const { setLoading } = useLoading();
+  const router = useRouter();
 
-  const handleSaveProduct = () => {};
-  const handleListProduct = async (listProductsParams: AxiosRequestConfig) => {
+  const handleSaveProduct = async (data: ProductSaveData, config: AxiosRequestConfig) => {
+    setLoading(true);
     try {
-      const { data } = await api.get("/tecadi/treinamento/produto", listProductsParams);
+      const { status } = await api.post("/tecadi/treinamento/produto", data, config);
+      if (status === 401) {
+        localStorage.clear();
+      }
+    } catch (error) {
+      const { response } = error as AxiosError;
+      if (response && response.status === 401) {
+        localStorage.clear();
+        router.push("/");
+      }
+    }
+    setLoading(false);
+  };
+  const handleListProduct = async (listProductsParams: AxiosRequestConfig) => {
+    setLoading(true);
+
+    try {
+      const { data, status } = await api.get("/tecadi/treinamento/produto", listProductsParams);
+      if (status === 400) {
+        setProducts([]);
+      }
       setProducts(data.list);
     } catch (error) {
-      localStorage.clear();
+      const { response } = error as AxiosError;
+      if (response && response.status === 400) {
+        setProducts([]);
+      }
+      if (response && response.status === 401) {
+        localStorage.clear();
+        router.push("/");
+      }
     }
+    setLoading(false);
   };
   const handleGetProduct = () => {};
   const handleUpdateProduct = () => {};
   const handleRemoveProduct = () => {};
-  return <ProductContext.Provider value={{ products, setProducts, handleListProduct }}>{children}</ProductContext.Provider>;
+  return <ProductContext.Provider value={{ products, setProducts, handleListProduct, handleSaveProduct }}>{children}</ProductContext.Provider>;
 };
 
 export const useProducts = () => useContext(ProductContext);
